chore: move BKTEvaluationDetails to a new file
